working queue example

+++ Frames.ReelConnector/Ast/ArithmeticNodes.cs
@@ -12,7 +12,7 @@ public class AdditionAstElement : BaseAstElement
 
         return left switch
         {
-            long l when right is long r => l + r,
+            double l when right is double r => l + r,
             int l when right is int r => l + r,
             double l when right is double r => l + r,
             string l when right is string r => l + r,
@@ -27,10 +27,10 @@ public class SubtractAstElement : BaseAstElement
     {
         var left = Evaluate(tree.Left, stateJson, bag);
         var right = Evaluate(tree.Right, stateJson, bag);
-
+        
         return left switch
         {
-            long l when right is long r => l - r,
+            double l when right is double r => l - r,
             int l when right is int r => l - r,
             double l when right is double r => l - r,
             _ => throw new InvalidOperationException("Incompatible types for subtraction")
@@ -47,7 +47,7 @@ public class MultiplyAstElement : BaseAstElement
 
         return left switch
         {
-            long l when right is long r => l * r,
+            double l when right is double r => l * r,
             int l when right is int r => l * r,
             double l when right is double r => l * r,
             _ => throw new InvalidOperationException("Incompatible types for multiplication")
@@ -64,7 +64,7 @@ public class DivideAstElement : BaseAstElement
 
         return left switch
         {
-            long l when right is long r => l / r,
+            double l when right is double r => l / r,
             int l when right is int r => l / r,
             double l when right is double r => l / r,
             _ => throw new InvalidOperationException("Incompatible types for multiplication")
